feat: unificando request e command
